requirements.txt
aiohttp==3.8.1
aioresponses==0.7.3
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
distlib==0.3.4
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
filelock==3.6.0
frozenlist==1.3.0
httpretty==1.1.4
idna==3.3
ipaddress==1.0.23
ipdata==3.4.3
lxml==4.8.0
multidict==6.0.2
Pillow==9.1.0
pipenv==2022.3.28
platformdirs==2.5.1
python-Wappalyzer==0.3.1
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
tqdm==4.64.0
urllib3==1.26.9
virtualenv==20.14.0
virtualenv-clone==0.5.7
wappalyzer==1.0
yarl==1.7.2
aiohttp==3.8.1
aioresponses==0.7.3
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.4
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
filelock==3.6.0
frozenlist==1.3.0
httpretty==1.1.4
idna==3.3
ipaddress==1.0.23
ipdata==3.4.3
kombu==5.2.4
lxml==4.8.0
multidict==6.0.2
Pillow==9.1.0
pipenv==2022.3.28
platformdirs==2.5.1
prompt-toolkit==3.0.29
pycparser==2.21
python-Wappalyzer==0.3.1
pytz==2022.1
rabbitmq==0.2.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
tqdm==4.64.0
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.0
virtualenv-clone==0.5.7
wappalyzer==1.0
wcwidth==0.2.5
yarl==1.7.2
